*Huge update* - Session runtime tracking + channel runtime accumulation - Way better support for multiple upstream playlists - Upstream connection limits are now implemented - Improved admin fields and views - Improved import processes - Automatically disable channels of unwanted channel types (e.g. ".m3u8") upon first import to avoid redirects - Better support for special characters - Improved logging - Passthru of response headers - Support for #EXTVLCOPT's http-user-agent and http-referrer - Probably more;-)
Known problems:
- EPG download probably still does'nt trigger correctly

+++ manager/lib/icon_helper.py
@@ -92,7 +92,7 @@ class IconHelper():
             
     def delete_icon_file(self):
         logger.info(f'ICON: Received delete request for {self.icon.url}') 
-        if os.path.exists(self.icon_filepath):
+        if (os.path.exists(self.icon_filepath)) and not (os.path.isdir(self.icon_filepath)):
             os.unlink(self.icon_filepath)
             logger.info(f'ICON: Icon deleted: {self.icon.url}')
         else:
